{"schema":"decodex/commit/1","summary":"Extract native host session surfaces","authority":"manual"}
